+++ src/pcl_seg_d435i/include/pcl_seg/processPointClouds.h
@@ -45,7 +45,7 @@ public:
 
     std::pair<PtCdtr<PointT>, PtCdtr<PointT>>RansacSegmentPlane(PtCdtr<PointT> cloud, int maxIterations, float distanceTol);
 
-    std::pair<PtCdtr<PointT>, PtCdtr<PointT>>RansacSegmentPlaneWithPose(PtCdtr<PointT> cloud, int maxIterations, float distanceTol, Eigen::Vector3d pose_data);
+    std::pair<PtCdtr<PointT>, PtCdtr<PointT>>RansacSegmentPlaneWithPose(PtCdtr<PointT> cloud, int maxIterations, float distanceTol, Eigen::Vector3f pose_data);
     
     std::vector<PtCdtr<PointT>>EuclideanClustering(PtCdtr<PointT> cloud, float clusterTolerance, int minSize,int maxSize);
